AMC-TXT: New global option ShowGroupText
When this option is enabled (default), the text after group start/end
marks i.e., after "*(" or "*)", is shown in the printed tests. This is
the same behavior as before this commit.

By setting this option to 0, no group marks are generated. This is
useful if groups are used merely as containers for similar questions
and have numquestions=1 option. In this case, it makes little sense to
add group text before and after every question.